+++ src/afl-fuzz-state.c
@@ -486,15 +486,15 @@ void read_afl_environment(afl_state_t *afl, char **envp) {
 
                               afl_environment_variable_len)) {
 
-            afl->min_length = atoi(
-                (u8 *)get_afl_env(afl_environment_variables[i]));
+            afl->min_length =
+                atoi((u8 *)get_afl_env(afl_environment_variables[i]));
 
           } else if (!strncmp(env, "AFL_INPUT_LEN_MAX",
 
                               afl_environment_variable_len)) {
 
-            afl->max_length = atoi(
-                (u8 *)get_afl_env(afl_environment_variables[i]));
+            afl->max_length =
+                atoi((u8 *)get_afl_env(afl_environment_variables[i]));
 
           }